linux-packages: facilitate modules out-of-core repo
Adapted from `kernel` package from core.

This allows the creation of kmod packages from the kernel that is in
OpenWrt core.

Some of these packages may be needed/used by packages in this feed, and may
not be of interest to OpenWrt core.

Having those kmods packaged here, helps to better decouple from core.

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>